Refactor subsidy constants into dedicated modules
