+++ zebra-chain/src/parameters/network/subsidy.rs
@@ -1,4 +1,4 @@
-//! Constants and calculations for Block Subsidy and Funding Streams
+//! Calculations for Block Subsidy and Funding Streams
 //!
 //! This module contains the consensus parameters which are required for
 //! verification.
@@ -12,46 +12,36 @@
 //! Typically, consensus parameters are accessed via a function that takes a
 //! `Network` and `block::Height`.
 
+pub mod constants;
+
 use std::collections::HashMap;
 
 use lazy_static::lazy_static;
 
 use crate::{
-    amount::{self, Amount, NonNegative, COIN},
+    amount::{self, Amount, NonNegative},
     block::{Height, HeightDiff},
     parameters::{constants::activation_heights, Network, NetworkUpgrade},
     transparent,
 };
 
-/// The largest block subsidy, used before the first halving.
-///
-/// We use `25 / 2` instead of `12.5`, so that we can calculate the correct value without using floating-point.
-/// This calculation is exact, because COIN is divisible by 2, and the division is done last.
-pub const MAX_BLOCK_SUBSIDY: u64 = ((25 * COIN) / 2) as u64;
-
-/// Used as a multiplier to get the new halving interval after Blossom.
-///
-/// Calculated as `PRE_BLOSSOM_POW_TARGET_SPACING / POST_BLOSSOM_POW_TARGET_SPACING`
-/// in the Zcash specification.
-pub const BLOSSOM_POW_TARGET_SPACING_RATIO: u32 = 2;
-
-/// Halving is at about every 4 years, before Blossom block time is 150 seconds.
-///
-/// `(60 * 60 * 24 * 365 * 4) / 150 = 840960`
-pub const PRE_BLOSSOM_HALVING_INTERVAL: HeightDiff = 840_000;
-
-/// After Blossom the block time is reduced to 75 seconds but halving period should remain around 4 years.
-pub const POST_BLOSSOM_HALVING_INTERVAL: HeightDiff =
-    PRE_BLOSSOM_HALVING_INTERVAL * (BLOSSOM_POW_TARGET_SPACING_RATIO as HeightDiff);
-
-/// The first halving height in the testnet is at block height `1_116_000`
-/// as specified in [protocol specification §7.10.1][7.10.1]
-///
-/// [7.10.1]: https://zips.z.cash/protocol/protocol.pdf#zip214fundingstreams
-pub(crate) const FIRST_HALVING_TESTNET: Height = Height(1_116_000);
-
-/// The first halving height in the regtest is at block height `287`.
-const FIRST_HALVING_REGTEST: Height = Height(287);
+use constants::{
+    mainnet::{
+        FUNDING_STREAM_ECC_ADDRESSES_MAINNET, FUNDING_STREAM_MG_ADDRESSES_MAINNET,
+        FUNDING_STREAM_ZF_ADDRESSES_MAINNET, POST_NU6_1_FUNDING_STREAM_FPF_ADDRESSES_MAINNET,
+        POST_NU6_FUNDING_STREAM_FPF_ADDRESSES_MAINNET, POST_NU6_FUNDING_STREAM_START_RANGE_MAINNET,
+    },
+    regtest::FIRST_HALVING_REGTEST,
+    testnet::{
+        FIRST_HALVING_TESTNET, FUNDING_STREAM_ECC_ADDRESSES_TESTNET,
+        FUNDING_STREAM_MG_ADDRESSES_TESTNET, FUNDING_STREAM_ZF_ADDRESSES_TESTNET,
+        POST_NU6_1_FUNDING_STREAM_FPF_ADDRESSES_TESTNET,
+        POST_NU6_FUNDING_STREAM_FPF_ADDRESSES_TESTNET, POST_NU6_FUNDING_STREAM_START_RANGE_TESTNET,
+    },
+    BLOSSOM_POW_TARGET_SPACING_RATIO, FUNDING_STREAM_RECEIVER_DENOMINATOR,
+    FUNDING_STREAM_SPECIFICATION, LOCKBOX_SPECIFICATION, MAX_BLOCK_SUBSIDY,
+    POST_BLOSSOM_HALVING_INTERVAL, PRE_BLOSSOM_HALVING_INTERVAL,
+};
 
 /// The funding stream receiver categories.
 #[derive(Serialize, Deserialize, Clone, Copy, Debug, Eq, Hash, PartialEq)]
@@ -106,21 +96,6 @@ impl FundingStreamReceiver {
         matches!(self, Self::Deferred)
     }
 }
-
-/// Denominator as described in [protocol specification §7.10.1][7.10.1].
-///
-/// [7.10.1]: https://zips.z.cash/protocol/protocol.pdf#zip214fundingstreams
-pub const FUNDING_STREAM_RECEIVER_DENOMINATOR: u64 = 100;
-
-/// The specification for pre-NU6 funding stream receivers, a URL that links to [ZIP-214].
-///
-/// [ZIP-214]: https://zips.z.cash/zip-0214
-pub const FUNDING_STREAM_SPECIFICATION: &str = "https://zips.z.cash/zip-0214";
-
-/// The specification for post-NU6 funding stream and lockbox receivers, a URL that links to [ZIP-1015].
-///
-/// [ZIP-1015]: https://zips.z.cash/zip-1015
-pub const LOCKBOX_SPECIFICATION: &str = "https://zips.z.cash/zip-1015";
 
 /// Funding stream recipients and height ranges.
 #[derive(Deserialize, Clone, Debug, Eq, PartialEq)]
@@ -369,121 +344,6 @@ lazy_static! {
     ];
 }
 
-/// The start height of post-NU6 funding streams on Mainnet as described in [ZIP-1015](https://zips.z.cash/zip-1015).
-const POST_NU6_FUNDING_STREAM_START_HEIGHT_MAINNET: u32 = 2_726_400;
-
-/// The start height of post-NU6 funding streams on Testnet as described in [ZIP-1015](https://zips.z.cash/zip-1015).
-const POST_NU6_FUNDING_STREAM_START_HEIGHT_TESTNET: u32 = 2_976_000;
-
-/// The one-time lockbox disbursement output addresses and amounts expected in the NU6.1 activation block's
-/// coinbase transaction on Mainnet.
-/// See:
-/// - <https://zips.z.cash/zip-0271#one-timelockboxdisbursement>
-/// - <https://zips.z.cash/zip-0214#mainnet-recipients-for-revision-2>
-pub const NU6_1_LOCKBOX_DISBURSEMENTS_MAINNET: [(&str, Amount<NonNegative>); 10] = [(
-    "t3ev37Q2uL1sfTsiJQJiWJoFzQpDhmnUwYo",
-    EXPECTED_NU6_1_LOCKBOX_DISBURSEMENTS_TOTAL_MAINNET.div_exact(10),
-); 10];
-
-/// The one-time lockbox disbursement output addresses and amounts expected in the NU6.1 activation block's
-/// coinbase transaction on Testnet.
-/// See:
-/// - <https://zips.z.cash/zip-0271#one-timelockboxdisbursement>
-/// - <https://zips.z.cash/zip-0214#testnet-recipients-for-revision-2>
-pub const NU6_1_LOCKBOX_DISBURSEMENTS_TESTNET: [(&str, Amount<NonNegative>); 10] = [(
-    "t2RnBRiqrN1nW4ecZs1Fj3WWjNdnSs4kiX8",
-    EXPECTED_NU6_1_LOCKBOX_DISBURSEMENTS_TOTAL_TESTNET.div_exact(10),
-); 10];
-
-/// The expected total amount of the one-time lockbox disbursement on Mainnet.
-/// See: <https://zips.z.cash/zip-0271#one-timelockboxdisbursement>.
-pub(crate) const EXPECTED_NU6_1_LOCKBOX_DISBURSEMENTS_TOTAL_MAINNET: Amount<NonNegative> =
-    Amount::new_from_zec(78_750);
-
-/// The expected total amount of the one-time lockbox disbursement on Testnet.
-/// See <https://zips.z.cash/zip-0271#one-timelockboxdisbursement>.
-pub(crate) const EXPECTED_NU6_1_LOCKBOX_DISBURSEMENTS_TOTAL_TESTNET: Amount<NonNegative> =
-    Amount::new_from_zec(78_750);
-
-/// The number of blocks contained in the post-NU6 funding streams height ranges on Mainnet or Testnet, as specified
-/// in [ZIP-1015](https://zips.z.cash/zip-1015).
-const POST_NU6_FUNDING_STREAM_NUM_BLOCKS: u32 = 420_000;
-
-/// The post-NU6 funding stream height range on Mainnet
-const POST_NU6_FUNDING_STREAM_START_RANGE_MAINNET: std::ops::Range<Height> =
-    Height(POST_NU6_FUNDING_STREAM_START_HEIGHT_MAINNET)
-        ..Height(POST_NU6_FUNDING_STREAM_START_HEIGHT_MAINNET + POST_NU6_FUNDING_STREAM_NUM_BLOCKS);
-
-/// The post-NU6 funding stream height range on Testnet
-const POST_NU6_FUNDING_STREAM_START_RANGE_TESTNET: std::ops::Range<Height> =
-    Height(POST_NU6_FUNDING_STREAM_START_HEIGHT_TESTNET)
-        ..Height(POST_NU6_FUNDING_STREAM_START_HEIGHT_TESTNET + POST_NU6_FUNDING_STREAM_NUM_BLOCKS);
-
-/// Address change interval function here as a constant
-/// as described in [protocol specification §7.10.1][7.10.1].
-///
-/// [7.10.1]: https://zips.z.cash/protocol/protocol.pdf#zip214fundingstreams
-pub const FUNDING_STREAM_ADDRESS_CHANGE_INTERVAL: HeightDiff = POST_BLOSSOM_HALVING_INTERVAL / 48;
-
-/// Number of addresses for each funding stream in the Mainnet.
-/// In the spec ([protocol specification §7.10][7.10]) this is defined as: `fs.addressindex(fs.endheight - 1)`
-/// however we know this value beforehand so we prefer to make it a constant instead.
-///
-/// [7.10]: https://zips.z.cash/protocol/protocol.pdf#fundingstreams
-pub const FUNDING_STREAMS_NUM_ADDRESSES_MAINNET: usize = 48;
-
-/// List of addresses for the ECC funding stream in the Mainnet.
-pub const FUNDING_STREAM_ECC_ADDRESSES_MAINNET: [&str; FUNDING_STREAMS_NUM_ADDRESSES_MAINNET] = [
-    "t3LmX1cxWPPPqL4TZHx42HU3U5ghbFjRiif",
-    "t3Toxk1vJQ6UjWQ42tUJz2rV2feUWkpbTDs",
-    "t3ZBdBe4iokmsjdhMuwkxEdqMCFN16YxKe6",
-    "t3ZuaJziLM8xZ32rjDUzVjVtyYdDSz8GLWB",
-    "t3bAtYWa4bi8VrtvqySxnbr5uqcG9czQGTZ",
-    "t3dktADfb5Rmxncpe1HS5BRS5Gcj7MZWYBi",
-    "t3hgskquvKKoCtvxw86yN7q8bzwRxNgUZmc",
-    "t3R1VrLzwcxAZzkX4mX3KGbWpNsgtYtMntj",
-    "t3ff6fhemqPMVujD3AQurxRxTdvS1pPSaa2",
-    "t3cEUQFG3KYnFG6qYhPxSNgGi3HDjUPwC3J",
-    "t3WR9F5U4QvUFqqx9zFmwT6xFqduqRRXnaa",
-    "t3PYc1LWngrdUrJJbHkYPCKvJuvJjcm85Ch",
-    "t3bgkjiUeatWNkhxY3cWyLbTxKksAfk561R",
-    "t3Z5rrR8zahxUpZ8itmCKhMSfxiKjUp5Dk5",
-    "t3PU1j7YW3fJ67jUbkGhSRto8qK2qXCUiW3",
-    "t3S3yaT7EwNLaFZCamfsxxKwamQW2aRGEkh",
-    "t3eutXKJ9tEaPSxZpmowhzKhPfJvmtwTEZK",
-    "t3gbTb7brxLdVVghSPSd3ycGxzHbUpukeDm",
-    "t3UCKW2LrHFqPMQFEbZn6FpjqnhAAbfpMYR",
-    "t3NyHsrnYbqaySoQqEQRyTWkjvM2PLkU7Uu",
-    "t3QEFL6acxuZwiXtW3YvV6njDVGjJ1qeaRo",
-    "t3PdBRr2S1XTDzrV8bnZkXF3SJcrzHWe1wj",
-    "t3ZWyRPpWRo23pKxTLtWsnfEKeq9T4XPxKM",
-    "t3he6QytKCTydhpztykFsSsb9PmBT5JBZLi",
-    "t3VWxWDsLb2TURNEP6tA1ZSeQzUmPKFNxRY",
-    "t3NmWLvZkbciNAipauzsFRMxoZGqmtJksbz",
-    "t3cKr4YxVPvPBG1mCvzaoTTdBNokohsRJ8n",
-    "t3T3smGZn6BoSFXWWXa1RaoQdcyaFjMfuYK",
-    "t3gkDUe9Gm4GGpjMk86TiJZqhztBVMiUSSA",
-    "t3eretuBeBXFHe5jAqeSpUS1cpxVh51fAeb",
-    "t3dN8g9zi2UGJdixGe9txeSxeofLS9t3yFQ",
-    "t3S799pq9sYBFwccRecoTJ3SvQXRHPrHqvx",
-    "t3fhYnv1S5dXwau7GED3c1XErzt4n4vDxmf",
-    "t3cmE3vsBc5xfDJKXXZdpydCPSdZqt6AcNi",
-    "t3h5fPdjJVHaH4HwynYDM5BB3J7uQaoUwKi",
-    "t3Ma35c68BgRX8sdLDJ6WR1PCrKiWHG4Da9",
-    "t3LokMKPL1J8rkJZvVpfuH7dLu6oUWqZKQK",
-    "t3WFFGbEbhJWnASZxVLw2iTJBZfJGGX73mM",
-    "t3L8GLEsUn4QHNaRYcX3EGyXmQ8kjpT1zTa",
-    "t3PgfByBhaBSkH8uq4nYJ9ZBX4NhGCJBVYm",
-    "t3WecsqKDhWXD4JAgBVcnaCC2itzyNZhJrv",
-    "t3ZG9cSfopnsMQupKW5v9sTotjcP5P6RTbn",
-    "t3hC1Ywb5zDwUYYV8LwhvF5rZ6m49jxXSG5",
-    "t3VgMqDL15ZcyQDeqBsBW3W6rzfftrWP2yB",
-    "t3LC94Y6BwLoDtBoK2NuewaEbnko1zvR9rm",
-    "t3cWCUZJR3GtALaTcatrrpNJ3MGbMFVLRwQ",
-    "t3YYF4rPLVxDcF9hHFsXyc5Yq1TFfbojCY6",
-    "t3XHAGxRP2FNfhAjxGjxbrQPYtQQjc3RCQD",
-];
-
 /// Functionality specific to block subsidy-related consensus rules
 pub trait ParameterSubsidy {
     /// Returns the minimum height after the first halving
@@ -547,135 +407,6 @@ impl ParameterSubsidy for Network {
         self.post_blossom_halving_interval() / 48
     }
 }
-
-/// List of addresses for the Zcash Foundation funding stream in the Mainnet.
-pub const FUNDING_STREAM_ZF_ADDRESSES_MAINNET: [&str; FUNDING_STREAMS_NUM_ADDRESSES_MAINNET] =
-    ["t3dvVE3SQEi7kqNzwrfNePxZ1d4hUyztBA1"; FUNDING_STREAMS_NUM_ADDRESSES_MAINNET];
-
-/// List of addresses for the Major Grants funding stream in the Mainnet.
-pub const FUNDING_STREAM_MG_ADDRESSES_MAINNET: [&str; FUNDING_STREAMS_NUM_ADDRESSES_MAINNET] =
-    ["t3XyYW8yBFRuMnfvm5KLGFbEVz25kckZXym"; FUNDING_STREAMS_NUM_ADDRESSES_MAINNET];
-
-/// Number of addresses for each post-NU6 funding stream on Mainnet.
-/// In the spec ([protocol specification §7.10][7.10]) this is defined as: `fs.addressindex(fs.endheight - 1)`
-/// however we know this value beforehand so we prefer to make it a constant instead.
-///
-/// [7.10]: https://zips.z.cash/protocol/protocol.pdf#fundingstreams
-pub const POST_NU6_FUNDING_STREAMS_NUM_ADDRESSES_MAINNET: usize = 12;
-
-/// List of addresses for the Major Grants post-NU6 funding stream on Mainnet administered by the Financial Privacy Fund (FPF).
-pub const POST_NU6_FUNDING_STREAM_FPF_ADDRESSES_MAINNET: [&str;
-    POST_NU6_FUNDING_STREAMS_NUM_ADDRESSES_MAINNET] =
-    ["t3cFfPt1Bcvgez9ZbMBFWeZsskxTkPzGCow"; POST_NU6_FUNDING_STREAMS_NUM_ADDRESSES_MAINNET];
-
-/// Number of addresses for each post-NU6.1 funding stream on Mainnet.
-/// In the spec ([protocol specification §7.10][7.10]) this is defined as: `fs.addressindex(fs.endheight - 1)`
-/// however we know this value beforehand so we prefer to make it a constant instead.
-///
-/// [7.10]: https://zips.z.cash/protocol/protocol.pdf#fundingstreams
-pub const POST_NU6_1_FUNDING_STREAMS_NUM_ADDRESSES_MAINNET: usize = 36;
-
-/// List of addresses for the Major Grants post-NU6.1 funding stream on Mainnet administered by the Financial Privacy Fund (FPF).
-pub const POST_NU6_1_FUNDING_STREAM_FPF_ADDRESSES_MAINNET: [&str;
-    POST_NU6_1_FUNDING_STREAMS_NUM_ADDRESSES_MAINNET] =
-    ["t3cFfPt1Bcvgez9ZbMBFWeZsskxTkPzGCow"; POST_NU6_1_FUNDING_STREAMS_NUM_ADDRESSES_MAINNET];
-
-/// Number of addresses for each funding stream in the Testnet.
-/// In the spec ([protocol specification §7.10][7.10]) this is defined as: `fs.addressindex(fs.endheight - 1)`
-/// however we know this value beforehand so we prefer to make it a constant instead.
-///
-/// [7.10]: https://zips.z.cash/protocol/protocol.pdf#fundingstreams
-pub const FUNDING_STREAMS_NUM_ADDRESSES_TESTNET: usize = 51;
-
-/// List of addresses for the ECC funding stream in the Testnet.
-pub const FUNDING_STREAM_ECC_ADDRESSES_TESTNET: [&str; FUNDING_STREAMS_NUM_ADDRESSES_TESTNET] = [
-    "t26ovBdKAJLtrvBsE2QGF4nqBkEuptuPFZz",
-    "t26ovBdKAJLtrvBsE2QGF4nqBkEuptuPFZz",
-    "t26ovBdKAJLtrvBsE2QGF4nqBkEuptuPFZz",
-    "t26ovBdKAJLtrvBsE2QGF4nqBkEuptuPFZz",
-    "t2NNHrgPpE388atmWSF4DxAb3xAoW5Yp45M",
-    "t2VMN28itPyMeMHBEd9Z1hm6YLkQcGA1Wwe",
-    "t2CHa1TtdfUV8UYhNm7oxbzRyfr8616BYh2",
-    "t2F77xtr28U96Z2bC53ZEdTnQSUAyDuoa67",
-    "t2ARrzhbgcpoVBDPivUuj6PzXzDkTBPqfcT",
-    "t278aQ8XbvFR15mecRguiJDQQVRNnkU8kJw",
-    "t2Dp1BGnZsrTXZoEWLyjHmg3EPvmwBnPDGB",
-    "t2KzeqXgf4ju33hiSqCuKDb8iHjPCjMq9iL",
-    "t2Nyxqv1BiWY1eUSiuxVw36oveawYuo18tr",
-    "t2DKFk5JRsVoiuinK8Ti6eM4Yp7v8BbfTyH",
-    "t2CUaBca4k1x36SC4q8Nc8eBoqkMpF3CaLg",
-    "t296SiKL7L5wvFmEdMxVLz1oYgd6fTfcbZj",
-    "t29fBCFbhgsjL3XYEZ1yk1TUh7eTusB6dPg",
-    "t2FGofLJXa419A76Gpf5ncxQB4gQXiQMXjK",
-    "t2ExfrnRVnRiXDvxerQ8nZbcUQvNvAJA6Qu",
-    "t28JUffLp47eKPRHKvwSPzX27i9ow8LSXHx",
-    "t2JXWPtrtyL861rFWMZVtm3yfgxAf4H7uPA",
-    "t2QdgbJoWfYHgyvEDEZBjHmgkr9yNJff3Hi",
-    "t2QW43nkco8r32ZGRN6iw6eSzyDjkMwCV3n",
-    "t2DgYDXMJTYLwNcxighQ9RCgPxMVATRcUdC",
-    "t2Bop7dg33HGZx3wunnQzi2R2ntfpjuti3M",
-    "t2HVeEwovcLq9RstAbYkqngXNEsCe2vjJh9",
-    "t2HxbP5keQSx7p592zWQ5bJ5GrMmGDsV2Xa",
-    "t2TJzUg2matao3mztBRJoWnJY6ekUau6tPD",
-    "t29pMzxmo6wod25YhswcjKv3AFRNiBZHuhj",
-    "t2QBQMRiJKYjshJpE6RhbF7GLo51yE6d4wZ",
-    "t2F5RqnqguzZeiLtYHFx4yYfy6pDnut7tw5",
-    "t2CHvyZANE7XCtg8AhZnrcHCC7Ys1jJhK13",
-    "t2BRzpMdrGWZJ2upsaNQv6fSbkbTy7EitLo",
-    "t2BFixHGQMAWDY67LyTN514xRAB94iEjXp3",
-    "t2Uvz1iVPzBEWfQBH1p7NZJsFhD74tKaG8V",
-    "t2CmFDj5q6rJSRZeHf1SdrowinyMNcj438n",
-    "t2ErNvWEReTfPDBaNizjMPVssz66aVZh1hZ",
-    "t2GeJQ8wBUiHKDVzVM5ZtKfY5reCg7CnASs",
-    "t2L2eFtkKv1G6j55kLytKXTGuir4raAy3yr",
-    "t2EK2b87dpPazb7VvmEGc8iR6SJ289RywGL",
-    "t2DJ7RKeZJxdA4nZn8hRGXE8NUyTzjujph9",
-    "t2K1pXo4eByuWpKLkssyMLe8QKUbxnfFC3H",
-    "t2TB4mbSpuAcCWkH94Leb27FnRxo16AEHDg",
-    "t2Phx4gVL4YRnNsH3jM1M7jE4Fo329E66Na",
-    "t2VQZGmeNomN8c3USefeLL9nmU6M8x8CVzC",
-    "t2RicCvTVTY5y9JkreSRv3Xs8q2K67YxHLi",
-    "t2JrSLxTGc8wtPDe9hwbaeUjCrCfc4iZnDD",
-    "t2Uh9Au1PDDSw117sAbGivKREkmMxVC5tZo",
-    "t2FDwoJKLeEBMTy3oP7RLQ1Fihhvz49a3Bv",
-    "t2FY18mrgtb7QLeHA8ShnxLXuW8cNQ2n1v8",
-    "t2L15TkDYum7dnQRBqfvWdRe8Yw3jVy9z7g",
-];
-
-/// List of addresses for the Zcash Foundation funding stream in the Testnet.
-pub const FUNDING_STREAM_ZF_ADDRESSES_TESTNET: [&str; FUNDING_STREAMS_NUM_ADDRESSES_TESTNET] =
-    ["t27eWDgjFYJGVXmzrXeVjnb5J3uXDM9xH9v"; FUNDING_STREAMS_NUM_ADDRESSES_TESTNET];
-
-/// List of addresses for the Major Grants funding stream in the Testnet.
-pub const FUNDING_STREAM_MG_ADDRESSES_TESTNET: [&str; FUNDING_STREAMS_NUM_ADDRESSES_TESTNET] =
-    ["t2Gvxv2uNM7hbbACjNox4H6DjByoKZ2Fa3P"; FUNDING_STREAMS_NUM_ADDRESSES_TESTNET];
-
-/// Number of addresses for each post-NU6 funding stream in the Testnet.
-/// In the spec ([protocol specification §7.10][7.10]) this is defined as: `fs.addressindex(fs.endheight - 1)`
-/// however we know this value beforehand so we prefer to make it a constant instead.
-///
-/// [7.10]: https://zips.z.cash/protocol/protocol.pdf#fundingstreams
-pub const POST_NU6_FUNDING_STREAMS_NUM_ADDRESSES_TESTNET: usize = 13;
-
-/// Number of addresses for each post-NU6 funding stream in the Testnet.
-/// In the spec ([protocol specification §7.10][7.10]) this is defined as: `fs.addressindex(fs.endheight - 1)`
-/// however we know this value beforehand so we prefer to make it a constant instead.
-///
-/// There are 27 funding stream periods across the 939,500 blocks for which the post-NU6.1 funding streams are
-/// active. See Testnet funding streams in revision 2 of <https://zips.z.cash/zip-0214#funding-streams>.
-///
-/// [7.10]: https://zips.z.cash/protocol/protocol.pdf#fundingstreams
-pub const POST_NU6_1_FUNDING_STREAMS_NUM_ADDRESSES_TESTNET: usize = 27;
-
-/// List of addresses for the Major Grants post-NU6 funding stream on Testnet administered by the Financial Privacy Fund (FPF).
-pub const POST_NU6_FUNDING_STREAM_FPF_ADDRESSES_TESTNET: [&str;
-    POST_NU6_FUNDING_STREAMS_NUM_ADDRESSES_TESTNET] =
-    ["t2HifwjUj9uyxr9bknR8LFuQbc98c3vkXtu"; POST_NU6_FUNDING_STREAMS_NUM_ADDRESSES_TESTNET];
-
-/// List of addresses for the Major Grants post-NU6.1 funding stream on Testnet administered by the Financial Privacy Fund (FPF).
-pub const POST_NU6_1_FUNDING_STREAM_FPF_ADDRESSES_TESTNET: [&str;
-    POST_NU6_1_FUNDING_STREAMS_NUM_ADDRESSES_TESTNET] =
-    ["t2HifwjUj9uyxr9bknR8LFuQbc98c3vkXtu"; POST_NU6_1_FUNDING_STREAMS_NUM_ADDRESSES_TESTNET];
 
 /// Returns the address change period
 /// as described in [protocol specification §7.10][7.10]
