+++ zebra-chain/src/parameters/network/subsidy/constants/regtest.rs
@@ -1,0 +1,6 @@
+//! Regtest-specific constants for block subsidies.
+
+use crate::block::Height;
+
+/// The first halving height in the regtest is at block height `287`.
+pub(crate) const FIRST_HALVING_REGTEST: Height = Height(287);
